+++ packages/pluggableWidgets/pie-doughnut-chart-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed an issue on pie chart onClick when multiple points are added to the same item.
+
 ## [5.1.0] - 2024-10-28
 
 ### Changed